- Changed used of int to fixed width interget types in cstdint - Added Debug Tracking allocation wrapper that can be used to wrap an Iallocator and perform tracking of its allocations - Fixed bug with deleting arrays of alliged pointers - Fixed some bugs with premake geeration - Renamed testproject to UnitTestPtoject - Added bat file to test code coverage of libary on unit test project

+++ Libs/Core/Include/Debug/Debug.h
@@ -28,7 +28,7 @@ namespace Flourish { namespace Debug
 	// Walks the stack and converts it to a string using symbols is found to get function Names.
 	// numFramesToOmit sets how many stack frames to skip in the output
 	// Return value indicated if the stacktrace was obtained ok and if it was truncated
-	StackTraceResult GetStackTrace(int numFramesToOmit,	char* callstackBufOut, int callstackBufOutLength );
+	StackTraceResult GetStackTrace(int32_t numFramesToOmit,	char* callstackBufOut, size_t callstackBufOutLength );
 }}
 
 //Include inline impimentation
